cleanup and add tests for 0.2.0
